moved stats.misc
