+++ skbio/distance/_base.py
@@ -14,10 +14,9 @@ import pandas as pd
 from scipy.spatial.distance import squareform
 
 from skbio._base import SkbioObject
-from skbio.stats._misc import _pprint_strs
 from skbio.util import find_duplicates
 from skbio.util._decorator import classonlymethod
-from skbio.util._misc import resolve_key
+from skbio.util._misc import resolve_key, _pprint_strs
 from skbio.util._plotting import PlottableMixin
 
 from ._utils import is_symmetric_and_hollow
