+++ skbio/ordination/_ordination_results.py
@@ -11,7 +11,7 @@ import functools
 import numpy as np
 
 from skbio._base import SkbioObject
-from skbio.stats._misc import _pprint_strs
+from skbio.util._misc import _pprint_strs
 from skbio.util._plotting import PlottableMixin
