+++ skbio/util/tests/test_misc.py
@@ -12,7 +12,7 @@ import unittest
 import numpy as np
 
 from skbio.util import cardinal_to_ordinal, safe_md5, find_duplicates, get_rng
-from skbio.util._misc import MiniRegistry, chunk_str, resolve_key
+from skbio.util._misc import MiniRegistry, chunk_str, resolve_key, _pprint_strs
 
 
 class TestMiniRegistry(unittest.TestCase):
@@ -273,6 +273,48 @@ class TestGetRng(unittest.TestCase):
         with self.assertRaises(ValueError) as cm:
             get_rng('hello')
         self.assertEqual(str(cm.exception), msg)
+
+
+class PPrintStrsTests(unittest.TestCase):
+    def test_truncation(self):
+        # truncation between items (on comma)
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=4)
+        self.assertEqual(obs, exp)
+
+        # truncation between items (on space)
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=5)
+        self.assertEqual(obs, exp)
+
+        # truncation on item
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=6)
+        self.assertEqual(obs, exp)
+
+        # truncation (no items)
+        exp = "..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=2)
+        self.assertEqual(obs, exp)
+
+    def test_no_truncation(self):
+        exp = "'a'"
+        obs = _pprint_strs(['a'], max_chars=3)
+        self.assertEqual(obs, exp)
+
+        exp = "'a', 'b', 'c'"
+        obs = _pprint_strs(['a', 'b', 'c'])
+        self.assertEqual(obs, exp)
+
+        exp = "'a', 'b', 'c'"
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=13)
+        self.assertEqual(obs, exp)
+
+    def test_non_default_delimiter_and_suffix(self):
+        exp = "'abc','defg',...."
+        obs = _pprint_strs(['abc', 'defg', 'hi', 'jklmno'], max_chars=14,
+                           delimiter=',', suffix='....')
+        self.assertEqual(obs, exp)
 
 
 if __name__ == '__main__':
